Create AbstractResourceTest
